requirements.txt
dj-rest-auth==3.0.0
Django==4.1.5
django-allauth==0.52.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
